+++ Samples~/Scripts/IssueAndTesting/ObjectPickerWorkaround/ObjectPickerTest.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.IssueAndTesting.ObjectPickerWorkaround
+{
+    public class ObjectPickerTest : MonoBehaviour
+    {
+        public GameObject objDefaultPicker;
+
+        [TestObjectPicker]
+        public string objectPicker;
+    }
+}