+++ Samples~/Scripts/IssueAndTesting/ObjectPickerWorkaround/TestObjectPickerAttribute.cs
@@ -1,0 +1,11 @@
+using SaintsField.Interfaces;
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.IssueAndTesting.ObjectPickerWorkaround
+{
+    public class TestObjectPickerAttribute : PropertyAttribute, ISaintsAttribute
+    {
+        public SaintsAttributeType AttributeType => SaintsAttributeType.Other;
+        public string GroupBy => "";
+    }
+}