+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
 
+## 4.5.1 ##
+
+UI Toolkit: fix `OnValueChanged` broken
+
 ## 4.5.0 ##
 
 1.  UI Toolkit: when `SaintsEditor` enabled, list/array can automatically add `SaintsRow` if necessary
 2.  UI Toolkit: `ListDrawerSettings`, same as above. Fix incorrect indent for the foldout
 3.  UI Toolkit: fix an issue that `PropRange` gave an error when working with list/array
 4.  IMGUI: fix height issue in `Table`, [#201](https://github.com/TylerTemp/SaintsField/issues/201)
-5.  UI Toolkit: fix width changes in horizontal `Layout` 
+5.  UI Toolkit: fix width changes in horizontal `Layout`
 6.  UI Toolkit: fix `Expandable`, `AnimatorState` foldout out of space when using inside `Layout`
 7.  UI Toolkit: fix `UnityEvent` can not be drawn correctly under some circumstances
 8.  UI Toolkit: fix `ShowInInspector` can not correctly draw a field when using a parent class type